Split bundle parsing to new package
