+++ packages/bundle-parser/.babelrc
@@ -1,0 +1,14 @@
+// Babel config for Node
+// Compiles sources, gulpfile and tests
+{
+  "presets": [
+    ["env", {
+      "targets": { "node": 4 },
+      "exclude": [
+        // Node 4 supports all the generator features that we use so there is no need in Regenerator
+        "transform-regenerator"
+      ]
+    }]
+  ],
+  "plugins": []
+}
