+++ packages/bundle-parser/README.md
@@ -1,0 +1,1 @@
+Maps bundled output files to the modules based on the given webpack stats object.
