+++ packages/bundle-parser/package.json
@@ -1,0 +1,45 @@
+{
+  "name": "@webpack-bundle-analyzer/bundle-parser",
+  "version": "3.0.0-alpha.1",
+  "description": "Maps webpack bundled output files to the modules based on the given webpack stats object",
+  "author": "Yury Grunin <grunin.ya@ya.ru>",
+  "contributors": [
+    "Yury Grunin <grunin.ya@ya.ru>",
+    "Vesa Laakso <laakso.vesa@gmail.com> (http://vesalaakso.com)"
+  ],
+  "license": "MIT",
+  "homepage": "https://github.com/webpack-contrib/webpack-bundle-analyzer",
+  "changelog": "https://github.com/webpack-contrib/webpack-bundle-analyzer/blob/master/CHANGELOG.md",
+  "bugs": {
+    "url": "https://github.com/webpack-contrib/webpack-bundle-analyzer/issues"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/webpack-contrib/webpack-bundle-analyzer.git"
+  },
+  "main": "src/index.js",
+  "engines": {
+    "node": ">= 4"
+  },
+  "scripts": {
+    "test": "mocha --exit --require babel-core/register",
+    "test-dev": "mocha --watch --require babel-core/register"
+  },
+  "files": [
+    "src"
+  ],
+  "dependencies": {
+    "acorn": "^5.3.0",
+    "lodash": "^4.17.4"
+  },
+  "devDependencies": {
+    "babel-core": "6.26.0",
+    "babel-preset-env": "1.6.1",
+    "mocha": "5.0.0"
+  },
+  "keywords": [
+    "webpack",
+    "bundle",
+    "analyzer"
+  ]
+}
