+++ packages/bundle-parser/src/index.js
@@ -1,0 +1,4 @@
+// TODO: Figure out a proper API
+module.exports = {
+  parseUtils: require('./parseUtils')
+};
