+++ packages/bundle-parser/src/parseUtils.js
