+++ packages/bundle-parser/test/.eslintrc.json
@@ -1,0 +1,13 @@
+{
+  "extends": "../../../.eslintrc.json",
+  "env": {
+    "mocha": true,
+    "browser": true
+  },
+  "globals": {
+    "sinon": true,
+    "expect": true,
+    "makeWebpackConfig": true,
+    "webpackCompile": true
+  }
+}
