+++ packages/bundle-parser/test/bundles/validBundleWithArrowFunction.js
