+++ packages/bundle-parser/test/bundles/validBundleWithArrowFunction.modules.json
