+++ packages/bundle-parser/test/bundles/validBundleWithEsNextFeatures.js
