+++ packages/bundle-parser/test/bundles/validBundleWithEsNextFeatures.modules.json
