+++ packages/bundle-parser/test/bundles/validCommonBundleWithDedupePlugin.js
