+++ packages/bundle-parser/test/bundles/validCommonBundleWithDedupePlugin.modules.json
