+++ packages/bundle-parser/test/bundles/validCommonBundleWithModulesAsArray.js
