+++ packages/bundle-parser/test/bundles/validCommonBundleWithModulesAsArray.modules.json
