+++ packages/bundle-parser/test/bundles/validCommonBundleWithModulesAsObject.js
