+++ packages/bundle-parser/test/bundles/validCommonBundleWithModulesAsObject.modules.json
