+++ packages/bundle-parser/test/bundles/validExtraBundleWithModulesAsArray.js
