+++ packages/bundle-parser/test/bundles/validExtraBundleWithModulesAsArray.modules.json
