+++ packages/bundle-parser/test/bundles/validExtraBundleWithModulesInsideArrayConcat.js
