+++ packages/bundle-parser/test/bundles/validExtraBundleWithModulesInsideArrayConcat.modules.json
