+++ packages/bundle-parser/test/bundles/validExtraBundleWithNamedChunk.modules.json
