+++ packages/bundle-parser/test/bundles/validJsonpWithArrayConcatAndEntryPoint.js
