+++ packages/bundle-parser/test/bundles/validJsonpWithArrayConcatAndEntryPoint.modules.json
