+++ packages/bundle-parser/test/bundles/validUmdLibraryBundleWithModulesAsArray.js
