+++ packages/bundle-parser/test/bundles/validUmdLibraryBundleWithModulesAsArray.modules.json
