+++ packages/plugin/package.json
@@ -29,8 +29,8 @@
     "src"
   ],
   "dependencies": {
+    "@webpack-bundle-analyzer/bundle-parser": "^3.0.0-alpha.1",
     "@webpack-bundle-analyzer/reporter-treemap": "^3.0.0-alpha.1",
-    "acorn": "^5.3.0",
     "bfj-node4": "^5.2.0",
     "chalk": "^2.3.0",
     "commander": "^2.13.0",
